requirements.txt
strands-agents>=0.1.0
a2a-sdk>=0.3.0
anthropic>=0.40.0
openai>=1.0.0
groq>=0.4.0
fastapi>=0.100.0
uvicorn>=0.30.0
aiosqlite>=0.20.0
httpx>=0.27.0
duckduckgo-search>=6.0.0
beautifulsoup4>=4.12.0
# Dome protection (used when DOME_ENABLED=1)
vijil-dome>=0.7.0
# OpenTelemetry (used for both Dome metrics and general observability)
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-http>=1.20.0
